Front end

+++ src/main/java/com/apepeskul/dao/PhoneDao.java
@@ -5,12 +5,17 @@ import com.apepeskul.domain.Phone;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
 @Transactional
 public interface PhoneDao {
-    public Phone get(int id);
-    public void delete (int id);
-    public void update (Phone phone);
-    public void save (Phone phone);
+    public Phone get(long id);
+
+    public void delete(long id);
+
+    public void update(Phone phone);
+
+    public void save(Phone phone);
+
     public List<Phone> getAll();
 
 }
